Flyway

+++ src/main/java/ru/example/bookmarket/exception/GenreNotFoundException.java
@@ -3,11 +3,12 @@ package ru.example.bookmarket.exception;
 import javax.persistence.EntityNotFoundException;
 
 public class GenreNotFoundException extends EntityNotFoundException {
-    public GenreNotFoundException(Long id) {
+    public GenreNotFoundException(int id) {
         super(String.format("Genre with id '%d' is not exist", id));
     }
 
     public GenreNotFoundException(String name) {
         super(String.format("Genre with name '%s' is not exist", name));
     }
+
 }
